Focus border (3/n): Settings UI + config template
- BordersSettings section on the General tab (grouped Form, reusing the shared
  NumberRow + ColorSwatchRow): enable toggle, color swatches, width + corner-
  radius (type-or-nudge), renderer picker (Overlay / Window server), and a
  motion-prediction toggle. Sub-controls disable when borders are off.
- SettingsModel border setters write the [borders] section via the surgical
  TOMLEditor (live-reloads the agent).
- config.toml template gains a documented [borders] section (#8), renumbered TOC.
- Fix: decode width/corner_radius as Int (TOMLKit is strict Int-vs-Double; the
  template + UI write integer px) and convert to Double — caught by the
  ConfigLoader/Validator real-config tests.

159 tests green; Borders section renders in the grouped-Form style.

+++ native/Sources/ZTSystem/ConfigLoader.swift
@@ -172,8 +172,8 @@ public enum ConfigLoader {
         var enabled: Bool?
         var backend: String?
         var color: String?
-        var width: Double?
-        var corner_radius: Double?
+        var width: Int?            // integer px (TOMLKit is strict Int-vs-Double)
+        var corner_radius: Int?
         var prediction: Bool?
     }
 
@@ -290,8 +290,8 @@ public enum ConfigLoader {
                 enabled: raw.borders?.enabled ?? false,
                 backend: raw.borders?.backend ?? "overlay",
                 color: raw.borders?.color ?? "blue",
-                width: raw.borders?.width ?? 4,
-                cornerRadius: raw.borders?.corner_radius ?? 9,
+                width: Double(raw.borders?.width ?? 4),
+                cornerRadius: Double(raw.borders?.corner_radius ?? 9),
                 prediction: raw.borders?.prediction ?? true))
     }
